layman_.py: Adds checks to ensure repo location
Previously it was made so that the initialization of the LaymanAPI
wouldn't occur if it already existed. This could however lead to
reporting the incorrect storage location for repositories reported
by repos.conf. Checks have been added to ensure that the LaymanAPI
will be re-initialized if the storage location reported by repos.conf
differs from what was already previously reported.